fixed oracles data

+++ src/interfaces/IBondOracle.sol
@@ -10,19 +10,19 @@ interface IBondOracle {
     error BondOracle__IncompleteResponse(uint256 length);
 
     event YieldUpdated(
-        uint64 twoYearYield,
-        uint64 fiveYearYield,
-        uint64 tenYearYield,
-        uint64 thirtyYearYield,
+        uint256 twoYearYield,
+        uint256 fiveYearYield,
+        uint256 tenYearYield,
+        uint256 thirtyYearYield,
         uint256 timestamp
     );
     event YieldUpdateFailed(bytes err);
 
     /// @notice Returns the yield for a given slot (1=2Y, 2=5Y, 3=10Y, 4=30Y)
     /// @dev Reverts if data is stale or slot is invalid
-    function getYield(uint256 slot) external view returns (uint64);
+    function getYield(uint256 slot) external view returns (uint256);
     function getAllYields() external view returns (BondYieldsResponse memory);
-    function updateYields(uint64[] memory values, uint256 timestamp, bytes memory err) external;
+    function updateYields(uint256[] memory values, uint256 timestamp, bytes memory err) external;
 
     /// @notice Returns true if the stored data is stale
     function isStale() external view returns (bool);
